+++ src/modules/powerdisplay/PowerDisplay/Helpers/SettingsDeepLink.cs
@@ -16,14 +16,14 @@ namespace PowerDisplay.Helpers
                 var directoryPath = System.AppContext.BaseDirectory;
                 if (mainExecutableIsOnTheParentFolder)
                 {
-                    // Need to go into parent folder for PowerToys.exe. Likely a WinUI3 App SDK application.
+                    // Need to go into parent folder for Kit.exe. Likely a WinUI3 App SDK application.
                     directoryPath = Path.Combine(directoryPath, "..");
-                    directoryPath = Path.Combine(directoryPath, "PowerToys.exe");
+                    directoryPath = Path.Combine(directoryPath, "Kit.exe");
                 }
                 else
                 {
-                    // PowerToys.exe is in the same path as the application.
-                    directoryPath = Path.Combine(directoryPath, "PowerToys.exe");
+                    // Kit.exe is in the same path as the application.
+                    directoryPath = Path.Combine(directoryPath, "Kit.exe");
                 }
 
                 Process.Start(new ProcessStartInfo(directoryPath) { Arguments = "--open-settings=PowerDisplay" });